# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "spelunk-server",
    "description": "Shared memory server for spelunk. Stores decisions, requirements, and context for a team and serves them over HTTP. Clients embed locally and send pre-computed vectors; the server stores and searches them.",
    "contact": {
      "name": "spelunk",
      "url": "https://github.com/usercise/spelunk"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Server liveness check. No authentication required.",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Server is up",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "ok"
              }
            }
          }
        }
      }
    },
    "/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List all projects registered on this server.",
        "operationId": "list_projects",
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "List memory entries for a project, optionally filtered by kind.",
        "operationId": "list_notes",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "Filter by kind (`decision`, `requirement`, `note`, `question`, `handoff`).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return (default: 20).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "archived",
            "in": "query",
            "description": "Include archived entries (default: false).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerNote"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Add a memory entry to a project. The project is auto-created on first write.",
        "description": "The client must supply a pre-computed `embedding` vector. All entries in\na project must use the same embedding dimension — the first write fixes it.",
        "operationId": "add_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug (e.g. `usercise/spelunk`)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Note created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Embedding dimension mismatch"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory/search": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Semantic search over memory entries using a pre-computed query embedding.",
        "operationId": "search_notes",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Nearest neighbours",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerNote"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory/{note_id}": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Get a single memory entry by ID.",
        "operationId": "get_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerNote"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "memory"
        ],
        "summary": "Delete a memory entry permanently.",
        "operationId": "delete_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoolResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory/{note_id}/archive": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Archive a memory entry. Archived entries are excluded from search and `ask`\ncontext but remain visible via `?archived=true`.",
        "operationId": "archive_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archive result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoolResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory/{note_id}/supersede": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Mark a memory entry as superseded by a newer one. The old entry is archived\nand linked to the new one.",
        "operationId": "supersede_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "description": "Note ID to supersede",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupersedeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Supersede result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoolResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Note not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/stats": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Return entry counts and embedding dimension for a project.",
        "operationId": "project_stats",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStats"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AddNoteRequest": {
        "type": "object",
        "required": [
          "kind",
          "title"
        ],
        "properties": {
          "body": {
            "type": "string"
          },
          "embedding": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Pre-computed embedding vector from the client (required for semantic search)."
          },
          "kind": {
            "type": "string",
            "description": "Kind of memory entry: `decision`, `requirement`, `note`, `question`, `handoff`."
          },
          "linked_files": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Source file paths this entry is linked to."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional tags for filtering."
          },
          "title": {
            "type": "string"
          }
        }
      },
      "AddNoteResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the created note."
          }
        }
      },
      "BoolResponse": {
        "type": "object",
        "required": [
          "changed"
        ],
        "properties": {
          "changed": {
            "type": "boolean",
            "description": "Whether the operation modified a record."
          }
        }
      },
      "CountResponse": {
        "type": "object",
        "required": [
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ListQuery": {
        "type": "object",
        "properties": {
          "archived": {
            "type": "boolean",
            "description": "Include archived entries (default: false)."
          },
          "kind": {
            "type": [
              "string",
              "null"
            ],
            "description": "Filter by kind (`decision`, `requirement`, `note`, `question`, `handoff`)."
          },
          "limit": {
            "type": "integer",
            "description": "Maximum number of results to return (default: 20).",
            "minimum": 0
          }
        }
      },
      "Project": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "embedding_dim",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of project creation."
          },
          "embedding_dim": {
            "type": "integer",
            "minimum": 0
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "slug": {
            "type": "string"
          }
        }
      },
      "ProjectStats": {
        "type": "object",
        "required": [
          "count",
          "total",
          "embedding_dim"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of active memory entries."
          },
          "embedding_dim": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total entries including archived."
          }
        }
      },
      "SearchRequest": {
        "type": "object",
        "required": [
          "embedding"
        ],
        "properties": {
          "embedding": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Query embedding vector (must match project's embedding dimension)."
          },
          "limit": {
            "type": "integer",
            "description": "Maximum number of results to return (default: 20).",
            "minimum": 0
          }
        }
      },
      "ServerNote": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "body",
          "tags",
          "linked_files",
          "created_at",
          "status"
        ],
        "properties": {
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation."
          },
          "distance": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Cosine distance from query (only present in search results)."
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kind": {
            "type": "string",
            "description": "Kind: `decision`, `requirement`, `note`, `question`, or `handoff`."
          },
          "linked_files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "`active` or `archived`."
          },
          "superseded_by": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
      "SupersedeRequest": {
        "type": "object",
        "required": [
          "new_id"
        ],
        "properties": {
          "new_id": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the new note that replaces the superseded one."
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Pass as `Authorization: Bearer <key>`. Not required when no key is configured on the server."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
  "tags": [
    {
      "name": "health",
      "description": "Liveness"
    },
    {
      "name": "projects",
      "description": "Project management"
    },
    {
      "name": "memory",
      "description": "Memory CRUD and semantic search"
    }
  ]
}